All config endpoints working. Working on updating all modules to trigger activeModule#ready and activeModule#setData methods

+++ webresources/js/config/groups.js
@@ -1,6 +1,85 @@
 var dataTable;
+var form;
 
 $(function(){
+	form = Blueprint.utils.FormBuilder.build("group-form");
+	form.readFieldsFromHTML();
+	form.trackListField($("#roles-list")[0]);
+	form.trackListField($("#group-members")[0]);
+	form.onsubmit(function(data, lists){
+		var requests = 0;
+		var errors = false;
+		if(data != undefined){
+			$.ajax({
+				method:			"post",
+				url:			"/rest/v1/config/directories/" + data.directory + "/groups/" + data.name,
+				data:			JSON.stringify(data),
+				contentType:	"application/json"
+			})
+			.done(updateLists)
+			.fail(function(xhr, status, err){
+				Blueprint.utils.Messaging.alert("An error occurred updating the group.", true, err);
+				errors = true;
+			});
+		}
+		else {
+			updateLists();
+		}
+		
+		function updateLists(){
+			var groupId = $("#name").val();
+			var directory = $("#directory").val();
+			requests += lists["roles-list"].added.length + lists["roles-list"].removed.length;
+			requests += lists["group-members"].added.length + lists["group-members"].removed.length;
+			
+			if(requests > 0){
+				for(var i=0;i<lists["roles-list"].added.length;i++){
+					modifyRoleMembership("post", lists["roles-list"].added[i], groupId, function(failed){
+						errors = errors || failed;
+						requests--;
+						checkComplete();
+					});
+				}
+				
+				for(var j=0;j<lists["roles-list"].removed.length;j++){
+					modifyRoleMembership("delete", lists["roles-list"].removed[j], groupId, function(failed){
+						errors = errors || failed;
+						requests--;
+						checkComplete();
+					});
+				}
+				
+				for(var k=0;k<lists["group-members"].added.length;k++){
+					modifyMember("post", lists["group-members"].added[k], directory, groupId, function(failed){
+						errors = errors || failed;
+						requests--;
+						checkComplete();
+					});
+				}
+				
+				for(var l=0;l<lists["group-members"].removed.length;l++){
+					modifyMember("delete", lists["group-members"].removed[l], directory, groupId, function(failed){
+						errors = errors || failed;
+						requests--;
+						checkComplete();
+					});
+				}
+			}
+			else {
+				checkComplete();
+			}
+		}
+		
+		function checkComplete(){
+			if(requests == 0){
+				Blueprint.utils.Messaging.alert("Group created/modified " + (errors ? "with errors." : "successfully."));
+				form.clearChangedState();
+				form.clear(function(){
+					clearForm(rebuildTable);
+				});
+			}
+		}
+	});
 	$("#group-config").addClass("hide");
 	$("#group-members-section").addClass("hide");
 	$("#add-group").on("click touch", function(evt){
@@ -15,7 +94,7 @@ $(function(){
 		buildRolesList(response);
 	})
 	.fail(function(xhr, status, err){
-		alert("Error retrieving groups data. Please consult the system logs for more information.\nError message: " + err);
+		Blueprint.utils.Messaging.alert("Error retrieving groups data.", true, err);
 	});
 	
 	$.ajax({
@@ -23,23 +102,31 @@ $(function(){
 		url:		"/rest/v1/config/directories",
 		dataType:	"json"
 	}).done(function(response){
-		$("#grp-directory").html(document.createElement("option"));
+		$("#directory").html(document.createElement("option"));
 		for(var key in response){
 			var option = document.createElement("option");
 			option.value = key;
 			option.text = key;
-			$("#grp-directory").append(option);
+			$("#directory").append(option);
 		}
 	})
 	.fail(function(xhr, status, err){
-		alert("Error retrieving directory data. Please consult the system logs for more information.\nError message: " + err);
+		Blueprint.utils.Messaging.alert("Error retrieving directory data.", true, err);
+	});
+	
+	$("#group-details-submit").off("click touch");
+	$("#group-details-submit").on("click touch", function(evt){
+		form.submit();
+	});
+	
+	$("#group-details-cancel").off("click touch");
+	$("#group-details-cancel").on("click touch", function(evt){
+		form.clear(clearForm);
 	});
 });
 
-
-
 function buildRolesList(groupData){
-	APP.utils.AdditiveList("roles-list");
+	Blueprint.utils.AdditiveList("roles-list");
 	$.ajax({
 		method:		"get",
 		url:		"/rest/v1/config/roles",
@@ -49,7 +136,7 @@ function buildRolesList(groupData){
 		buildGroupsTable(groupData);
 	})
 	.fail(function(xhr, status, err){
-		alert("Error retrieving roles data. Please consult the system logs for more information.\nError message: " + err);
+		Blueprint.utils.Messaging.alert("Error retrieving roles data.", true, err);
 	});
 }
 
@@ -78,25 +165,27 @@ function buildGroupsTable(groupsResponse){
 				title:			"",
 				className:		"delete",
 				data:			null,
-				defaultContent: "<span class='delete fa fa-times'/>",
+				defaultContent: "<span class='delete fa fa-times'></span>",
 				searchable:		false,
 				orderable:		false,
 				width:			"5%",
 				createdCell:	function(cell, cellData, rowData, rowIndex, cellIndex){
 					$(cell).find("span.delete").on("click touch", function(evt){
 						evt.stopPropagation();
-						var conf = confirm("Deleting a group cannot be undone. Proceed?");
-						
-						$.ajax({
-							method:		"delete",
-							url:		"/rest/v1/config/directories/" + rowData.directory + "/groups/" + rowData.name
-						})
-						.done(function(response){
-							alert("Group removed successfully.");
-							clearForm();
-						})
-						.fail(function(xhr, status, err){
-							alert("Error removing group. Please check the system logs for more information.\nError message: " + err);
+						Blueprint.utils.Messaging.confirm("Deleting a group cannot be undone. Proceed?", function(conf){
+							if(conf){
+								$.ajax({
+									method:		"delete",
+									url:		"/rest/v1/config/directories/" + rowData.directory + "/groups/" + rowData.name
+								})
+								.done(function(response){
+									Blueprint.utils.Messaging.alert("Group removed successfully.");
+									clearForm();
+								})
+								.fail(function(xhr, status, err){
+									Blueprint.utils.Messaging.alert("Error removing group.", true, err);
+								});
+							}
 						});
 					});
 				}
@@ -108,7 +197,7 @@ function buildGroupsTable(groupsResponse){
 		pagingType:		"first_last_numbers",
 		rowCallback:	function(row, data, index){
 			$(row).on("click touch", function(evt){
-				cancelForm(function(){
+				form.clear(function(){
 					$("#groups").find("tr.selected").removeClass("selected");
 					$(row).addClass("selected");
 				
@@ -122,7 +211,7 @@ function buildGroupsTable(groupsResponse){
 }
 
 function generateUsersList(){
-	APP.utils.AdditiveList("group-members");
+	Blueprint.utils.AdditiveList("group-members");
 	$.ajax({
 		method:		"get",
 		url:		"/rest/v1/config/users",
@@ -139,36 +228,35 @@ function generateUsersList(){
 		
 		$("#group-members")[0].setListOptions(users);
 	});
+	
+	Blueprint.modules.activeModule().ready();
 }
 
 function showGroupForm(groupDetails){
 	var groupConfigForm = document.getElementById("group-config");
-	groupConfigForm.groupDetails = groupDetails;
 	$(groupConfigForm).removeClass("hide");
 	$("#add-group").addClass("hide");
-	$("#grp-directory").removeAttr("disabled");
-	$("#grp-displayName").off("keyup");
+	$("#directory").removeAttr("disabled");
+	$("#displayName").off("keyup");
 	$("#group-members-section").removeClass("hide");
 	$("#roles-list")[0].clear();
 	$("#group-members")[0].clear();
-	groupConfigForm.roles = [];
-	groupConfigForm.members = [];
 	
 	if(groupDetails != undefined){
-		$("#grp-directory").val(groupDetails.directory).attr("disabled", true);
-		$("#grp-displayName").val(groupDetails.displayName);
-		$("#grp-name").val(groupDetails.name);
+		$("#directory").val(groupDetails.directory).attr("disabled", true);
+		$("#displayName").val(groupDetails.displayName);
+		$("#name").val(groupDetails.name);
 		
 		$.ajax({
 			method:		"get",
-			url:		"/rest/v1/auth/groups/" + groupDetails.name + "/roles",
+			url:		"/rest/v1/config/groups/" + groupDetails.name + "/roles",
 			dataType:	"json"
 		})
 		.done(function(response){
 			for(var i=0;i<response.length;i++){
-				groupConfigForm.roles.push(response[i].roleId);
 				$("#roles-list")[0].addOption(response[i].roleId);
 			}
+			$("#roles-list")[0].setOriginalState();
 		});
 		
 		$.ajax({
@@ -180,132 +268,17 @@ function showGroupForm(groupDetails){
 			for(var i=0;i<response.length;i++){
 				$("#group-members")[0].addOption(response[i]);
 			}
-			
-			groupConfigForm.members = response;
+			$("#group-members")[0].setOriginalState();
 		});
 	}
 	else {
-		$("#grp-displayName").on("keyup", function(evt){
-			var value = $("#grp-displayName").val().replace(" ", "-").toLowerCase();
-			$("#grp-name").val(value);
+		$("#displayName").on("keyup", function(evt){
+			var value = $("#displayName").val().replace(" ", "-").toLowerCase();
+			$("#name").val(value);
 		});
 	}
 	
-	$("#group-details-submit").off("click touch");
-	$("#group-details-submit").on("click touch", function(evt){
-		evt.preventDefault();
-		submitData();
-	});
 	
-	$("#group-details-cancel").off("click touch");
-	$("#group-details-cancel").on("click touch", function(evt){
-		evt.preventDefault();
-		cancelForm();
-	});
-}
-
-function submitData(){
-	var groupConfigForm = document.getElementById("group-config");
-	var groupDetails = groupConfigForm.groupDetails;
-	var roles = groupConfigForm.roles || [];
-	var members = groupConfigForm.members || [];
-	var data = {};
-	var directory = $("#grp-directory").val();
-	var groupId = $("#grp-name").val();
-	var update = true;
-	var errors = false;
-	var changed = false;
-	
-	if(groupDetails == undefined){
-		update = false;
-		var missing = [];
-		$("div.input-label.required").each(function(index, el){
-			if($(el).parent().find(".input-field").val() == ""){
-				missing.push($(el).text());
-			}
-		});
-		
-		if(missing.length > 0){
-			alert("Could not complete the request. The following required fields do not have a value:\n\t" + missing.join("\n\t"));
-			return;
-		}
-	}
-	
-	groupDetails = groupDetails || {};
-	groupDetails.displayName = groupDetails.displayName || "";
-	
-	if($("#grp-displayName").val() != groupDetails.displayName){
-		data.displayName = $("#grp-displayName").val();
-	}
-	var requests = 0;
-	if(Object.keys(data).length > 0){
-		requests++;
-		changed = true;
-		$.ajax({
-			method:			"post",
-			url:			"/rest/v1/config/directories/" + directory + "/groups/" + groupId,
-			data:			JSON.stringify(data),
-			contentType:	"application/json"
-		})
-		.fail(function(xhr, status, err){
-			alert("An error occurred updating the group. Please check the system logs for more information.\nError message: " + err);
-			errors = true;
-		})
-		.always(function(){
-			requests--;
-			checkComplete();
-		});
-	}
-	
-	var rolesDifferences = $("#roles-list")[0].getDifferences(roles);
-	changed = changed || (rolesDifferences.added.length + rolesDifferences.removed.length) > 0;
-	requests += (rolesDifferences.added.length + rolesDifferences.removed.length);
-	for(var i=0;i<rolesDifferences.added.length;i++){
-		modifyRoleMembership("post", rolesDifferences.added[i], groupId, function(err){
-			errors = errors || err;
-			requests--;
-			checkComplete();
-		});
-	}
-	
-	for(var j=0;j<rolesDifferences.removed.length;j++){
-		modifyRoleMembership("delete", rolesDifferences.removed[j], groupId, function(err){
-			errors = errors || err;
-			requests--;
-			checkComplete();
-		});
-	}
-	
-	var memberDifferences = $("#group-members")[0].getDifferences(members);
-	changed = changed || (memberDifferences.added.length + memberDifferences.removed.length) > 0;
-	requests += (memberDifferences.added.length + memberDifferences.removed.length);
-	for(var k=0;k<memberDifferences.added.length;k++){
-		modifyMember("post", memberDifferences.added[k], directory, groupId, function(err){
-			errors = errors || err;
-			requests--;
-			checkComplete();
-		});
-	}
-	
-	for(var l=0;l<memberDifferences.removed.length;l++){
-		modifyMember("delete", memberDifferences.removed[l], directory, groupId, function(err){
-			errors = errors || err;
-			requests--;
-			checkComplete();
-		});
-	}
-	
-	checkComplete();
-	
-	function checkComplete(){
-		if(requests == 0 && changed){
-			alert("Group " + (update ? "updated": "created") + " " + (errors ? "with errors." : "successfully"));
-			clearForm(rebuildTable);
-		}
-		else if(!changed){
-			alert("No changes have been made yet.");
-		}
-	}
 }
 
 function modifyRoleMembership(method, role, groupId, next){
@@ -317,7 +290,7 @@ function modifyRoleMembership(method, role, groupId, next){
 		contentType:	"application/json"
 	})
 	.fail(function(xhr, status, err){
-		alert("Error " + ((method == "post") ? "attaching role '" + role.name + "' to " : "removing role '" + role.name + "' from ") + "group. Please check the system logs for more information.\nError message: " + err);
+		Blueprint.utils.Messaging.alert("Error " + ((method == "post") ? "attaching role '" + role.name + "' to " : "removing role '" + role.name + "' from ") + "group.", true, err);
 		error = true;
 	})
 	.always(function(){
@@ -334,48 +307,12 @@ function modifyMember(method, user, directory, groupId, next){
 		contentType:	"application/json"
 	})
 	.fail(function(xhr, status, err){
-		alert("Error " + ((method == "post") ? "adding user '" + user.fn + " " + user.ln + "' to " : "removing user '" + user.fn + " " + user.ln + "' from ") + "group. Please check the system logs for more information.\nError message: " + err);
+		Blueprint.utils.Messaging.alert("Error " + ((method == "post") ? "adding user '" + user.fn + " " + user.ln + "' to " : "removing user '" + user.fn + " " + user.ln + "' from ") + "group.", true, err);
 		error = true;
 	})
 	.always(function(){
 		next(error);
 	});
-}
-
-function cancelForm(next){
-	var groupConfigForm = document.getElementById("group-config");
-	var groupDetails = groupConfigForm.groupDetails;
-	var roles = groupConfigForm.roles || [];
-	var members = groupConfigForm.members || [];
-	var changed = false;
-	var cancel = true;
-	
-	if(groupDetails == undefined){
-		if($("#grp-directory").val() != "" || $("#grp-name").val() != ""){
-			changed = true;
-		}
-	}
-	
-	groupDetails = groupDetails || {};
-	groupDetails.displayName = groupDetails.displayName || "";
-	
-	if($("#grp-displayName").val() != groupDetails.displayName){
-		changed = true;
-	}
-	
-	var rolesDifferences = $("#roles-list")[0].getDifferences(roles);
-	changed = changed || (rolesDifferences.added.length + rolesDifferences.removed.length) > 0;
-	
-	var membersDifferences = $("#group-members")[0].getDifferences(members);
-	changed = changed || (membersDifferences.added.length + membersDifferences.removed.length) > 0;
-	
-	if(changed){
-		cancel = confirm("You have made changes that will be lost if you cancel. Proceed?");
-	}
-	
-	if(cancel){
-		clearForm(next);
-	}
 }
 
 function clearForm(next){
@@ -410,6 +347,27 @@ function rebuildTable(){
 		dataTable.draw();
 	})
 	.fail(function(xhr, status, err){
-		alert("Error retrieving groups data. Please consult the system logs for more information.\nError message: " + err);
+		Blueprint.utils.Messaging.alert("Error retrieving groups data.", true, err);
+	});
+}
+
+Blueprint.modules.activeModule().setData = function(data){
+	
+}
+
+Blueprint.modules.activeModule().destroy = function(next){
+	form.clear(function(){
+		delete dataTable;
+		delete form;
+		delete buildRolesList;
+		delete buildGroupsTable;
+		delete generateUsersList;
+		delete showGroupForm;
+		delete modifyRoleMembership;
+		delete modifyMember;
+		delete clearForm;
+		delete rebuildTable;
+		
+		next();
 	});
 }
